[refactor] config file 폴더 생성 및 이동
- config file을 한군데에 모아 관리가 편하도록 함
- client url은 환경 변수를 사용하도록 함
- 사용하지 않는 클래스 삭제
