+++ BE/src/main/java/cuketmon/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package cuketmon;
+package cuketmon.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
