+++ BE/src/main/java/cuketmon/config/WebClientConfig.java
@@ -1,4 +1,4 @@
-package cuketmon;
+package cuketmon.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
