+++ BE/src/main/java/cuketmon/config/WebSocketConfig.java
@@ -1,5 +1,6 @@
-package cuketmon.battle;
+package cuketmon.config;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
@@ -12,11 +13,13 @@ import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerCo
 @EnableWebSocket
 public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
 
+    @Value("${client.redirect-url}")
+    private String CLIENT_URL;
+
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry) {
         registry.addEndpoint("/ws")
-                .setAllowedOriginPatterns("chrome-extension://*", "http://localhost:3000",
-                        "https://frolicking-gnome-f1b1ad.netlify.app/make")
+                .setAllowedOriginPatterns("chrome-extension://*", "http://localhost:3000", CLIENT_URL)
                 .withSockJS();
     }
