--- BE/src/main/java/cuketmon/skill/dto/SkillDTO.java
@@ -1,4 +1,0 @@
-package cuketmon.skill.dto;
-
-public class SkillDTO {
-}
